Fix compatibility with ModLoader dispenser hook. Fixes a bug with IDispenseHandler too - though it means that interface is now deprecated

+++ fml/common/cpw/mods/fml/common/modloader/BaseModProxy.java
@@ -89,7 +89,7 @@ public interface BaseModProxy
 
     public abstract void onItemPickup(EntityPlayer player, ItemStack item);
 
-    public abstract int dispenseEntity(World world, ItemStack item, Random rnd, double x, double y, double z, int xVel, int zVel, double entX,
+    public abstract int dispenseEntity(World world, ItemStack item, Random rnd, int x, int y, int z, int xVel, int zVel, double entX,
             double entY, double entZ);
 
     public abstract void serverCustomPayload(NetServerHandler handler, Packet250CustomPayload packet);
